refactor(router): move urlParser into the rules folder
