+++ modules/angular2/src/router/rules/url_parser.ts
