+++ modules/angular2/test/router/path_recognizer_spec.ts
@@ -11,7 +11,7 @@ import {
 } from 'angular2/testing_internal';
 
 import {PathRecognizer} from 'angular2/src/router/rules/path_recognizer';
-import {parser, Url} from 'angular2/src/router/url_parser';
+import {parser, Url} from 'angular2/src/router/rules/url_parser';
 
 export function main() {
   describe('PathRecognizer', () => {
